GPU: Fixed Major Sprite Display Bug
Fixed Issue causing Sprites to be placed in very incorrect positions.
This was due to parenthesis being wrapped around operations that were
not intended to be a part of the readMemory address, but were meant to
subtract the value returned from the readMemory call.

This is now solved.

+++ Source/GameBeak-Frontend/Classes/GPU.cs
@@ -297,12 +297,12 @@ namespace GameBeak_Frontend.Classes
 
             for (int i = 0; i < 40; i++)
             {
-                y = Core.beakMemory.readMemory((ushort)((mapAddress + (i * 4)) - 16));
+                y = Core.beakMemory.readMemory((ushort)((mapAddress + (i * 4)))) - 16;
 
                 if (y > -8)
                 {
 
-                    x = Core.beakMemory.readMemory((ushort)((mapAddress + (i * 4) + 1) - 8));
+                    x = (byte)(Core.beakMemory.readMemory((ushort)((mapAddress + (i * 4) + 1))) - 8);
 
                     bool isSpriteOnLine = (y <= lineY) && ((y + ((spriteSize == false) ? 8 : 16)) > lineY);
                     //(y <= lineY)  if the start of the sprite is at lineY or before it
